Extract app from server.js
